hoist constants in more packages
