+++ src/wxyz_datagrid/src/wxyz/datagrid/js/src/index.ts
@@ -1,4 +1,4 @@
-import __package__ from '../package.json';
-
-export const NAME = __package__.name;
-export const VERSION = __package__.version;
+export * from './constants';
+export * from './widgets';
+export * from './widgets/models/cells';
+export * from './widgets/pmodels/jsonmodel';
