+++ src/wxyz_json_e/src/wxyz/json_e/js/package.json
@@ -48,6 +48,10 @@
       "@deathbeds/wxyz-core": {
         "bundled": false,
         "singleton": true
+      },
+      "codemirror": {
+        "bundled": false,
+        "singleton": true
       }
     }
   },
