+++ src/wxyz_lab/src/wxyz/lab/js/src/constants.ts
@@ -1,0 +1,4 @@
+import __package__ from '../package.json';
+
+export const NAME = __package__.name;
+export const VERSION = __package__.version;
