+++ src/wxyz_lab/src/wxyz/lab/js/src/index.ts
@@ -1,4 +1,2 @@
-import __package__ from '../package.json';
-
-export const NAME = __package__.name;
-export const VERSION = __package__.version;
+export * from './constants';
+export * from './widgets';
