+++ src/wxyz_svg/src/wxyz/svg/js/src/index.ts
@@ -1,4 +1,1 @@
-import __package__ from '../package.json';
-
-export const NAME = __package__.name;
-export const VERSION = __package__.version;
+export * from './constants';
